Renaming interface -> Interface to align with the class name.
Renaming ued_tilt_series -> ued for simplicity since this will be a user-called function.
