+++ pyTEM/lib/interface/mixins/AcquisitionMixin.py
