+++ pyTEM/lib/interface/mixins/BeamBlankerMixin.py
