+++ pyTEM/lib/interface/mixins/MagnificationMixin.py
