+++ pyTEM/lib/interface/mixins/ModeMixin.py
