+++ pyTEM/lib/interface/mixins/VacuumMixin.py
